add filtering for travel airplanes & create custom mixin for filtering airplanes
