+++ cargo/views.py
@@ -21,6 +21,7 @@ from cargo.serializers import (
     CargoFlightDetailSerializer,
     CargoOrderDetailSerializer
 )
+from config.custom.mixins import AirPlaneFilteringMixin
 
 
 class CargoViewSet(viewsets.ModelViewSet):
@@ -65,37 +66,9 @@ class CargoViewSet(viewsets.ModelViewSet):
         ),
     ]
 )
-class CargoAirplaneViewSet(viewsets.ModelViewSet):
+class CargoAirplaneViewSet(AirPlaneFilteringMixin, viewsets.ModelViewSet):
     queryset = CargoAirplane.objects.all()
     serializer_class = CargoAirplaneSerializer
-
-    def get_queryset(self):
-        queryset = self.queryset
-
-        min_range = self.request.query_params.get("min_range")
-        max_range = self.request.query_params.get("max_range")
-        model = self.request.query_params.get("model")
-        country = self.request.query_params.get("country")
-
-        if min_range:
-            queryset = queryset.filter(
-                max_range_km__gte=int(min_range),
-            )
-        if max_range:
-            queryset = queryset.filter(
-                max_range_km__lte=int(max_range),
-            )
-        if model:
-            queryset = queryset.filter(
-                model__icontains=model,
-            )
-        if country:
-            queryset = queryset.filter(
-                country_of_origin__icontains=country,
-            )
-
-        return queryset
-
 
 
 class CargoFlightViewSet(viewsets.ModelViewSet):
